Rename `allocator.rs` to `io.rs`
