+++ ostd/src/arch/x86/mod.rs
@@ -2,11 +2,11 @@
 
 //! Platform-specific code for the x86 platform.
 
-mod allocator;
 pub mod boot;
 pub(crate) mod cpu;
 pub mod device;
 pub(crate) mod ex_table;
+mod io;
 pub mod iommu;
 pub(crate) mod irq;
 pub(crate) mod kernel;
@@ -18,8 +18,8 @@ pub mod task;
 pub mod timer;
 pub mod trap;
 
-use allocator::{construct_io_mem_allocator_builder, construct_io_port_allocator_builder};
 use cfg_if::cfg_if;
+use io::{construct_io_mem_allocator_builder, construct_io_port_allocator_builder};
 use spin::Once;
 use x86::cpuid::{CpuId, FeatureInfo};
